plat/common: Add support to cache CPU id
Add support to cache the CPU id and retrieve it with low overhead.
The raw instruction to retrieve the CPU id might be expensive (this
is the case of CPUID when executed in VM environment on x86). This
commit provides a caching mechanism for the x86 architecture based
on the TSC_AUX MSR.

Signed-off-by: Federico Parola <federico.parola@polito.it>